Fix the test suite to work properly with Mocha.
The test used `this` to store all kinds of things that worked in Mocha 1.13 but
stoppped working somewhere around Mocha 1.19. It was a really bad idea to use
`this` in the first place. Fixed now.

Clean up the code that was generated from CoffeeScript.

Fix some tests that became stricter in later versions of
Sinon. calledWithExactly, for instance would silently convert numbers to
strings, etc.