We should focus the primary button rather than the accept button.
For one thing, focusing the accept bootprompt-accept button is a breaking change
with previously formally released versions. Secondly, the de facto standard for
dialogs is to mark the default action with btn-primary. This is the action that
should be performed if the user does not click a specific button presented at
the bottom of the dialog box.