Don't use undefined as a return value for callbacks.
We want callbacks to be able to return a boolean, undefined, or return nothing,
which boils down to returning an undefined value. ``boolean | undefined`` makes
the TS compiler require the function to use ``return undefined`` when not
returnig anything would be fine. So we use ``boolean | void`` instead which does
allow not returning anything and allows ``return undefined``.