refactor: simplify handling of overloads
alert, confirm and prompt are overloaded to allow calling with an option object,
or with a string and a callback. mergeDialogOptions used to handle the
variations, but this made mergeDialogOptions hard to read. And it obscured what
actually happened with the code. The code has been refactored to make the
crystal clear what happens, and simplify mergeDialogOptions.