refactor: default ``onEscape`` on ``dialog`` changed
The default was that if ``onEscape`` was not set, then the dialog would not
close when Escape was pressed. The new default is that the dialog closes when
the key is pressed.

Rationales:

1. The previous behavior was inconsistent with the rest of Bootprompt. Not
setting ``onEscape`` on the specialized functions (``alert``, ``confirm``,
``prompt``) meant close the modal when Escape is pressed. Only ``dialog`` had a
default that was the exact opposite. This violates the principle of least
surprise.

2. The Bootstrap default is to have modals that are closed when Escape is
pressed. Bootprompt should not *silently* modify the default behavior of
Bootstrap.

These two problems were inherited from Bootbox when Bootprompt was forked.

BREAKING CHANGE: the meaning an unset ``onEscape`` on ``dialog`` is the exact
opposite of what it used to be.