fix: ensure that modals are removed
At some point the code of the hidden.bs.modal handler was set to do nothing if
the event target was not equal to this. There was a comment with the code saying
that it was probably no longer necessary because Bootstrap added namespacing.

The namespacing does take care of most problematic situations. There could still
be problems if someone embeds a modal in a modal but a) that's pushing it and b)
that's probably not even supported by Bootstrap. There could also be the odd
namespace clash but I don't seem value worring about this without evidence that
it happens.

Conversely the code as it existed prior to this commit was problematic because
the handler was added with ``.one``. If the handler got an event that it decided
not to act on, then it would do nothing but would be removed. If the correct
event was later produced the handler was no longer there to handle it!! And the
modal would never be removed.

It would have been possible to alter the logic to remove the handler only when
the event was the right one, but it seems simpler to just remove the event test.