fix: improve type inference for the prompt$ function
Previously, the typing of ``prompt$`` was such that it sucked at determining
what the promise it returns would resolve to. It tried to infer it from
``callback`` but that field was not always specified, and this would mess up the
inference.

We now use an entirely different approach for inferring the value.